moved Service into ChurchCRM dir
